// src/app/core/enums/ChannelGroup.ts
 export enum ChannelGroup
    {
        Twitter = 1,
        Facebook = 2,
        Instagram = 3,
        Youtube = 4,
        LinkedIn = 5,
        GooglePlus = 6,
        GooglePlayStore = 7,
        AutomotiveIndia = 8,
        Blogs = 9,
        Booking = 10,
        ComplaintWebsites = 11,
        CustomerCare = 12,
        DiscussionForums = 13,
        ECommerceWebsites = 14,
        Expedia = 15,
        HolidayIQ = 16,
        MakeMyTrip = 17,
        MyGov = 18,
        News = 19,
        ReviewWebsites = 20,
        TeamBHP = 21,
        TripAdvisor = 22,
        Videos = 23,
        Zomato = 24,
        Email = 25, //
        GoogleMyReview = 26, //
        ElectronicMedia = 27,
        WhatsApp = 28, //
        WebsiteChatBot = 29,
        AppStoreReviews = 30
    }