// src/app/core/enums/ChannelType.ts
export enum ChannelType {
    PublicTweets = 1,
    FBPageUser = 2,
    FBPublic = 3,
    Blogs = 4,
    DiscussionForums = 5,
    News = 6,
    Twitterbrandmention = 7,
    FBComments = 8,
    TripAdvisor = 9,
    BrandTweets = 10,
    FbMediaPosts = 11,
    FBMediaComments = 12,
    TeamBHPPosts = 13,
    TeamBHPComments = 14,
    TeamBHPOtherPostsComments = 15,
    ComplaintWebsites = 16,
    DirectMessages = 17,
    YouTubePosts = 18,
    YouTubeComments = 19,
    InstagramPagePosts = 20,
    InstagramUserPosts = 21,
    InstagramComments = 22,
    InstagramPublicPosts = 23,
    GooglePagePosts = 24,
    GoogleUserPosts = 25,
    GooglePublicPosts = 26,
    GoogleComments = 27,
    CustomerCare = 28,
    Expedia = 29,
    Booking = 30,
    ReviewWebsitePosts = 31,
    ReviewWebsiteComments = 32,
    ECommercePosts = 33,
    ECommerceComments = 34,
    HolidayIQReview = 37,
    ZomatoPost = 38,
    ZomatoComment = 39,
    FBMessages = 40,
    Videos = 41,
    GooglePlayStore = 42,
    LinkedInPageUser = 43,
    LinkedInPublic = 44,
    LinkedInComments = 45,
    LinkedInMediaPosts = 46,
    LinkedInMediaComments = 47,
    LinkedinMessage = 48,
    FBReviews = 49,
    AutomotiveIndiaPost = 50,
    AutomotiveIndiaComment = 51,
    AutomotiveIndiaOtherPostsComments = 52,
    MakeMyTrip = 53,
    Email = 54,
    GoogleMyReview = 55,
    ElectronicMedia = 56,
    GMBQuestion = 58,
    GMBAnswers = 59,
    WhatsApp = 60,
    FacebookChatbot = 61,
    WesiteChatbot = 62,
    AndroidChatbot = 63,
    IOSChatbot = 64,
    LineChatbot = 65,
    WhatsAppChatbot = 66,
    AppStoreReviews = 67,
    DiscourseTopic = 68,
    DiscoursePost = 69
}